requirements.txt
amqp==5.0.2
asgiref==3.3.1
astroid==2.4.2
billiard==3.6.3.0
boto3==1.16.30
botocore==1.19.30
celery==5.0.3
celery-progress==0.0.14
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
dj-database-url==0.5.0
Django==3.1.4
django-celery-results==2.0.0
django-ckeditor==6.0.0
django-crispy-forms==1.10.0
django-embed-video==1.3.3
django-environ==0.4.5
django-heroku==0.3.1
django-js-asset==1.2.2
django-paypal==1.0.0
django-storages==1.10.1
gunicorn==20.0.4
idna==2.10
isort==5.6.4
jmespath==0.10.0
kombu==5.0.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.0.1
prompt-toolkit==3.0.8
psycopg2==2.8.6
psycopg2-binary==2.8.6
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.4
redis==3.5.3
requests==2.25.0
s3transfer==0.3.3
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
wrapt==1.12.1
